Add: `extensions` resolvers
