+++ src/data/resolvers/colony.ts
@@ -10,10 +10,6 @@ import {
   extensions,
   getExtensionHash,
   ColonyClientV5,
-  getEvents,
-  getLogs,
-  getBlockTime,
-  ROOT_DOMAIN_ID,
 } from '@colony/colony-js';
 
 import ENS from '~lib/ENS';
@@ -40,7 +36,6 @@ import { COLONY_TOTAL_BALANCE_DOMAIN_ID } from '~constants';
 import { createAddress } from '~utils/web3';
 import { log } from '~utils/debug';
 import { Color } from '~core/ColorTag';
-import extensionData from '~data/staticData/extensionData';
 
 import { getToken } from './token';
 import {
@@ -601,84 +596,6 @@ export const colonyResolvers = ({
         },
         [],
       );
-    },
-  },
-  ColonyExtension: {
-    async details({ address, extensionId }, { colonyAddress }) {
-      const extension = extensionData[extensionId];
-      const colonyClient = await colonyManager.getClient(
-        ClientType.ColonyClient,
-        colonyAddress,
-      );
-      if (colonyClient.clientVersion === ColonyVersion.GoerliGlider) {
-        throw new Error('Colony version too old');
-      }
-
-      const { neededColonyPermissions } = extension;
-
-      const missingPermissions = await Promise.resolve(
-        neededColonyPermissions.reduce(async (roles, role) => {
-          const hasRole = await colonyClient.hasUserRole(
-            address,
-            ROOT_DOMAIN_ID,
-            role,
-          );
-          if (!hasRole) return [...(await roles), role];
-          return roles;
-        }, Promise.resolve([])),
-      );
-
-      const installFilter = networkClient.filters.ExtensionInstalled(
-        getExtensionHash(extensionId),
-        colonyAddress,
-        null,
-      );
-      const installLogs = await getLogs(networkClient, installFilter);
-      let installedBy = AddressZero;
-      let installedAt = 0;
-
-      if (
-        installLogs[0] &&
-        installLogs[0].transactionHash &&
-        installLogs[0].blockHash
-      ) {
-        const { blockHash, transactionHash } = installLogs[0];
-        const receipt = await networkClient.provider.getTransactionReceipt(
-          transactionHash,
-        );
-        installedBy = receipt.from || AddressZero;
-        const time = await getBlockTime(networkClient.provider, blockHash);
-        installedAt = time || 0;
-      }
-
-      const extensionClient = await colonyClient.getExtensionClient(
-        extensionId,
-      );
-
-      const deprecated = await extensionClient.getDeprecated();
-
-      // If no initializationParams are present it does not need initialization
-      // and will set to be true by default
-      let initialized = !extension.initializationParams;
-      if (!initialized) {
-        // Otherwise we look for the presence of an initialization event
-        // eslint-disable-next-line max-len
-        const initializedFilter = extensionClient.filters.ExtensionInitialised();
-        const initializedEvents = await getEvents(
-          extensionClient,
-          initializedFilter,
-        );
-        initialized = !!initializedEvents.length;
-      }
-
-      return {
-        __typename: 'ColonyExtensionDetails',
-        deprecated,
-        missingPermissions,
-        initialized,
-        installedBy,
-        installedAt,
-      };
     },
   },
 });
