+++ src/data/resolvers/extensions.ts
@@ -1,0 +1,122 @@
+import { Resolvers } from '@apollo/client';
+import { AddressZero } from 'ethers/constants';
+import {
+  ClientType,
+  ColonyVersion,
+  getExtensionHash,
+  getEvents,
+  getLogs,
+  getBlockTime,
+  ROOT_DOMAIN_ID,
+} from '@colony/colony-js';
+
+import { Context } from '~context/index';
+
+import extensionData from '~data/staticData/extensionData';
+
+export const extensionResolvers = ({
+  colonyManager: { networkClient },
+  colonyManager,
+}: Required<Context>): Resolvers => ({
+  Query: {
+    async networkExtensionVersion(_, { extensionId }) {
+      /*
+       * Prettier is being stupid again
+       */
+      // eslint-disable-next-line max-len
+      const extensionAddedToNetworkFilter = networkClient.filters.ExtensionAddedToNetwork(
+        getExtensionHash(extensionId),
+        null,
+      );
+      const extensionAddedEvents = await getEvents(
+        networkClient,
+        extensionAddedToNetworkFilter,
+      );
+      if (extensionAddedEvents.length) {
+        const latestEvent =
+          extensionAddedEvents[extensionAddedEvents.length - 1];
+        const version = latestEvent?.values?.version;
+        return version.toNumber();
+      }
+      return 0;
+    },
+  },
+  ColonyExtension: {
+    async details({ address, extensionId }, { colonyAddress }) {
+      const extension = extensionData[extensionId];
+      const colonyClient = await colonyManager.getClient(
+        ClientType.ColonyClient,
+        colonyAddress,
+      );
+      if (colonyClient.clientVersion === ColonyVersion.GoerliGlider) {
+        throw new Error('Colony version too old');
+      }
+
+      const { neededColonyPermissions } = extension;
+
+      const missingPermissions = await Promise.resolve(
+        neededColonyPermissions.reduce(async (roles, role) => {
+          const hasRole = await colonyClient.hasUserRole(
+            address,
+            ROOT_DOMAIN_ID,
+            role,
+          );
+          if (!hasRole) return [...(await roles), role];
+          return roles;
+        }, Promise.resolve([])),
+      );
+
+      const installFilter = networkClient.filters.ExtensionInstalled(
+        getExtensionHash(extensionId),
+        colonyAddress,
+        null,
+      );
+      const installLogs = await getLogs(networkClient, installFilter);
+      let installedBy = AddressZero;
+      let installedAt = 0;
+
+      if (
+        installLogs[0] &&
+        installLogs[0].transactionHash &&
+        installLogs[0].blockHash
+      ) {
+        const { blockHash, transactionHash } = installLogs[0];
+        const receipt = await networkClient.provider.getTransactionReceipt(
+          transactionHash,
+        );
+        installedBy = receipt.from || AddressZero;
+        const time = await getBlockTime(networkClient.provider, blockHash);
+        installedAt = time || 0;
+      }
+
+      const extensionClient = await colonyClient.getExtensionClient(
+        extensionId,
+      );
+
+      const deprecated = await extensionClient.getDeprecated();
+
+      // If no initializationParams are present it does not need initialization
+      // and will set to be true by default
+      let initialized = !extension.initializationParams;
+      if (!initialized) {
+        // Otherwise we look for the presence of an initialization event
+        // eslint-disable-next-line max-len
+        const initializedFilter = extensionClient.filters.ExtensionInitialised();
+        const initializedEvents = await getEvents(
+          extensionClient,
+          initializedFilter,
+        );
+        initialized = !!initializedEvents.length;
+      }
+
+      return {
+        __typename: 'ColonyExtensionDetails',
+        deprecated,
+        missingPermissions,
+        initialized,
+        installedBy,
+        installedAt,
+      };
+    },
+  },
+});
